% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aegean.R
\docType{data}
\name{IberianBirds}
\alias{IberianBirds}
\title{A SAR dataset describing *****}
\format{
A list with two elements: data frame with 2 columns
  and 90 rows. Each row contains the area of an island in the
  Aegean (1st column) and the number of inverts on that island
  (2nd column).
}
\source{
***
}
\usage{
data(IberianBirds)
}
\description{
A sample dataset in ***
}
\examples{
data(IberianBirds)
}
\keyword{datasets}
